+++ Changelog.md
@@ -22,6 +22,7 @@
 - New: Ignore Owner option for the Bouncing Betty and Mine
 - New: The Security Camera can now have a Lens, making it possible to tint the camera view in a different color
 - New: The Security Camera now supports the Smart Module, allowing the owner to set a default viewing direction
+- New: Electrified Iron Fences and Electrified Iron Fence Gates can now be pushed by Reinforced Pistons
 - Change: The Reinforced Hopper screen now shows "Reinforced Hopper" instead of "Item Hopper" as its default inventory title
 - Change: The hitbox of the Sentry has been changed, which means that only the base of Sentries can be interacted with now
 - Change: The recipe for the Electrified Iron Fence/Fence Gate now requires any wooden reinforced fence/fence gate